requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.2
django-import-export==2.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyexcel==0.7.0
pyexcel-io==0.6.6
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
texttable==1.6.4
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
